move flash-attn monkey patch alongside the others
